fix for issue #2803 - Added volttron-upgrade command to update both auth and move backup.sqlite files Added relevant documentation
