+++ volttron/platform/upgrade/move_sqlite_files.py
@@ -39,14 +39,14 @@
 import sys
 import shutil
 from pathlib import Path
+import glob
+import re
 from gevent import monkey as curious_george
 curious_george.patch_all(thread=False, select=False)
 
 from volttron.platform import get_home
 from volttron.platform.aip import AIPplatform
-from volttron.platform.auth import AuthFile
 from volttron.platform.instance_setup import fail_if_instance_running
-from volttron.platform.keystore import KeyStore
 
 
 def get_aip():
@@ -58,87 +58,39 @@ def get_aip():
     return aip
 
 
-def get_agent_path(agent_dir_path, agent_dir_suffix):
-    """
-    Stand-alone method based off of agent_name method from AIPplatform.
-    Gets the path to the agent file of the specified directory if it exists.
-    """
-    for agent_name in agent_dir_path.iterdir():
-        try:
-            for agent_subdir in agent_name.iterdir():
-                agent_dir = agent_name.joinpath(
-                    agent_subdir.stem + f".{agent_dir_suffix}")
-                if agent_dir.exists():
-                    return agent_dir
-        # Ignore files that are not directories
-        except NotADirectoryError:
-            pass
-
-    raise KeyError(agent_dir_path.stem)
-
-
-def upgrade_old_agents(aip):
+def move_historian_cache_files(aip):
     """
     Moves any keystore.json from agent-data to dist-info.
     Only applies to agents in auth file.
     """
 
     vhome = Path(aip.env.volttron_home)
-    agent_map = aip.get_agent_identity_to_uuid_mapping()
-
-    auth_file = AuthFile()
     install_dir = vhome.joinpath("agents")
-    for agent in agent_map:
-        agent_path = install_dir.joinpath(agent_map[agent])
-        try:
-            agent_data = get_agent_path(agent_path, 'agent-data')
-        # Skip if no agent-data exists
-        except KeyError as err:
-            print(f"agent-data not found for {err}")
-            continue
+    # pattern example - (/vhome/agents/uuid/agentname-version/)(backup.sqlite)
+    # pattern example - (vhome/agents/uuid/agentname-version/)(data/subdir/sqlitehistoriandb.sqlite)
+    pattern = "(" + str(install_dir) + "/[^/]*/[^/]*/)(.*)"
+    re_pattern = re.compile(pattern)
+    # Look for all .sqlite files in installed agents
+    print(f"Attempting to move backup.sqlite files in {install_dir} into corresponding agent-data directory")
+    # currently this is only used for backup.sqlite
+    # In 9.0 we could use the same code for *.sqlite files
+    # for example ones created by sqlitetagging, topic watcher, weather etc. when we make agent-data folder default
+    # agent write directory for all core agents
+    glob_path = str(install_dir.joinpath("**/backup.sqlite"))
+    for sqlite_file in glob.glob(glob_path, recursive=True):
+        result = re_pattern.match(sqlite_file).groups()
+        agent_dir = result[0]  # <volttron home>/agents/uuid/<agent name-version>
+        source_file = result[1].split('/', 1)[0]  # file or directory name to be moved to agent-data folder
+        source_path = str(Path(agent_dir).joinpath(source_file))
+        dest_dir = aip.get_agent_data_dir(agent_path=agent_dir)
+        if source_path != dest_dir:
+            # if file is not already in agent-data dir
+            result = shutil.move(source_path, dest_dir)
 
-        keystore_path = agent_data.joinpath('keystore.json')
-        try:
-            dist_info = get_agent_path(agent_path, 'dist-info')
-        # Skip if no dist-info exists
-        except KeyError as err:
-            print(f"dist-info not found for {err}")
-            continue
-        keystore_dest_path = dist_info.joinpath('keystore.json')
-
-        if keystore_path.exists():
-            agent_keystore = KeyStore(keystore_path)
-            for entry in auth_file.read()[0]:
-                # Only move if agent exists in auth file
-                if entry.credentials == agent_keystore.public:
-                    shutil.move(str(keystore_path), str(keystore_dest_path))
-                    break
-    return
-
-
-
-def get_identity_credentials(aip):
-    """Returns a dictionary containing a mapping from publickey to identity"""
-
-    agent_map = aip.get_agent_identity_to_uuid_mapping()
-    agent_credential_map = {}
-    for agent in agent_map:
-        agent_credential = aip.get_agent_keystore(agent_map[agent]).public
-        agent_credential_map[agent_credential] = agent
-    return agent_credential_map
-
-
-def set_auth_identities(agent_credential_map):
-    """Updates auth entries' identity field in auth file based on existing agents"""
-
-    auth_file = AuthFile()
-    entries, deny_entries, groups, roles = auth_file.read()
-    for entry in entries:
-        for credential in agent_credential_map:
-            if entry.credentials == credential:
-                entry.identity = agent_credential_map[credential]
-    auth_file._write(entries, deny_entries, groups, roles)
-    return
+            # print from uuid dir name so that it is easy to read
+            print_src = source_path.split(str(install_dir)+"/")[1]
+            print_dest = result.split(str(install_dir) + "/")[1]
+            print(f"Moved {print_src} to {print_dest}")
 
 
 def main():
@@ -146,10 +98,10 @@ def main():
 
     fail_if_instance_running()
     aip = get_aip()
-    upgrade_old_agents(aip)
-    identity_map = get_identity_credentials(aip)
-    set_auth_identities(identity_map)
-    print("Auth File Update Complete!")
+    move_historian_cache_files(aip)
+    print("Moving historian backup files complete. "
+          "You can now safely upgrade historian agents other than SQLITE Historian with vctl install --force. "
+          "If using using SQLite historian please back up and restore sqlite historian's db manually")
 
 
 def _main():
